Move device_stat to its own file
