+++ src/fairseq2/device.py
@@ -7,20 +7,29 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from collections.abc import Mapping
+from collections.abc import Iterator, Mapping
+from contextlib import contextmanager
 from typing import final
 
 import torch
-from typing_extensions import override
 
 from fairseq2.context import RuntimeContext
 from fairseq2.error import InternalError
-from fairseq2.typing import CPU, Device
+from fairseq2.typing import CPU, DataType, Device
+from fairseq2.utils.cuda import CudaContext, TorchCudaContext
 from fairseq2.utils.env import (
     InvalidEnvironmentVariableError,
     get_device_from_env,
     get_int_from_env,
 )
+
+
+def determine_default_device(context: RuntimeContext) -> Device:
+    cuda_context = TorchCudaContext()
+
+    device_accessor = DefaultDeviceAccessor(context.env, cuda_context)
+
+    return device_accessor.get()
 
 
 @final
@@ -117,93 +126,19 @@ class DeviceDetectionError(Exception):
     pass
 
 
-def determine_default_device(context: RuntimeContext) -> Device:
-    cuda_context = TorchCudaContext()
-
-    device_accessor = DefaultDeviceAccessor(context.env, cuda_context)
-
-    return device_accessor.get()
-
-
-class CudaContext(ABC):
-    @abstractmethod
-    def is_available(self) -> bool: ...
-
-    @abstractmethod
-    def device_count(self) -> int: ...
-
-    @abstractmethod
-    def set_default_device(self, device: Device) -> None: ...
-
-
-@final
-class TorchCudaContext(CudaContext):
-    @override
-    def is_available(self) -> bool:
-        return torch.cuda.is_available()
-
-    @override
-    def device_count(self) -> int:
-        return torch.cuda.device_count()
-
-    @override
-    def set_default_device(self, device: Device) -> None:
-        torch.cuda.set_device(device)
-
-
-class DeviceStatTracker(ABC):
-    @abstractmethod
-    def get_stats(self) -> dict[str, object]: ...
-
-    @abstractmethod
-    def reset(self) -> None: ...
-
-
-@final
-class NoopDeviceStatTracker(DeviceStatTracker):
-    @override
-    def get_stats(self) -> dict[str, object]:
-        return {}
-
-    @override
-    def reset(self) -> None:
-        pass
-
-
-@final
-class CudaDeviceStatTracker(DeviceStatTracker):
-    _device: Device
-    _total_memory: int
-
-    def __init__(self, device: Device) -> None:
-        self._device = device
-
-        props = torch.cuda.get_device_properties(device)
-
-        self._total_memory = props.total_memory
-
-    @override
-    def get_stats(self) -> dict[str, object]:
-        stats = torch.cuda.memory_stats(self._device)
-
-        peak_active_mem = stats["active_bytes.all.peak"]
-        peak_active_mem_ratio = peak_active_mem / self._total_memory
-
-        peak_reserved_mem = stats["reserved_bytes.all.peak"]
-        peak_reserved_mem_ratio = peak_reserved_mem / self._total_memory
-
-        return {
-            "peak_active_mem": peak_active_mem,
-            "peak_active_mem_ratio": peak_active_mem_ratio,
-            "peak_reserved_mem": peak_reserved_mem,
-            "peak_reserved_mem_ratio": peak_reserved_mem_ratio,
-        }
-
-    @override
-    def reset(self) -> None:
-        torch.cuda.reset_peak_memory_stats()
-
-
 class SupportsDeviceTransfer(ABC):
     @abstractmethod
     def to(self, device: Device) -> None: ...
+
+
+@contextmanager
+def default_device_and_dtype(device: Device, dtype: DataType) -> Iterator[None]:
+    original_dtype = torch.get_default_dtype()
+
+    torch.set_default_dtype(dtype)
+
+    try:
+        with device:
+            yield
+    finally:
+        torch.set_default_dtype(original_dtype)
