+++ src/fairseq2/recipes/_generator.py
@@ -15,7 +15,7 @@ from torch.profiler import record_function
 from typing_extensions import override
 
 from fairseq2.datasets import DataReader, DataReadError
-from fairseq2.device import DeviceStatTracker, SupportsDeviceTransfer
+from fairseq2.device import SupportsDeviceTransfer
 from fairseq2.error import InternalError, InvalidOperationError
 from fairseq2.gang import GangError, Gangs
 from fairseq2.logging import log
@@ -23,6 +23,7 @@ from fairseq2.metrics import MetricBag, MetricBagError
 from fairseq2.metrics.recorders import MetricRecorder, MetricRecordError
 from fairseq2.profilers import Profiler
 from fairseq2.typing import CPU, ContextManager, DataType
+from fairseq2.utils.device_stat import DeviceStatTracker
 from fairseq2.utils.progress import ProgressReporter, ProgressTask
 from fairseq2.utils.rng import RngBag
 from fairseq2.utils.stopwatch import Stopwatch
